# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ratings/{id}": {
      "post": {
        "operationId": "Ratings_addRating",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddedRatingResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Ratings"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "add Rating"
      }
    },
    "/ratings": {
      "get": {
        "operationId": "Ratings_getRatings",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "isWatchTogether",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "rate",
                "userRate"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "sortByUserId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "format": "int32",
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Ratings"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Ratings"
      }
    },
    "/ratings/movie/{id}": {
      "get": {
        "operationId": "Ratings_getOneRating",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatedMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Ratings"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get One Rating"
      }
    },
...
    "/rooms/{roomId}/invitations": {
      "post": {
        "operationId": "RoomInvitations_inviteUser",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserToRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationCreatedResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInvitations"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "invite User"
      },
      "get": {
        "operationId": "RoomInvitations_getRoomInvitations",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInvitations"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Room Invitations"
      }
    },
    "/invite-links/{token}": {
      "get": {
        "operationId": "PublicInviteLinks_getOneByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "PublicInviteLinks"
        ],
        "security": [],
        "summary": "get One By Token"
      }
    },
...
    "/rooms/{roomId}/invite-links": {
      "post": {
        "operationId": "RoomInviteLinks_create",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomInviteLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedInviteLinkResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInviteLinks"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "create"
      },
      "get": {
        "operationId": "RoomInviteLinks_findAll",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteLinksResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInviteLinks"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "find All"
      }
    },
    "/rooms/{roomId}/invite-links/{id}": {
      "patch": {
        "operationId": "RoomInviteLinks_update",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomInviteLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedInviteLinkResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInviteLinks"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "update"
      },
      "delete": {
        "operationId": "RoomInviteLinks_remove",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemovedInviteLinkResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "RoomInviteLinks"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "remove"
      }
    },
...
    "/rooms/{id}": {
      "get": {
        "operationId": "Rooms_getRoomById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Room By Id"
      }
    },
...
  },
  "info": {
    "title": "Paradiso API",
    "description": "Movie sharing, rooms, ratings and invitations. Successful responses include apiVersion. Use an access token with Authorize; refresh tokens are only for /auth/refresh-tokens.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddRatingDto": {
        "type": "object",
        "properties": {
          "rate": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 10
          },
          "movieId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "rate",
          "movieId"
        ],
        "additionalProperties": false
      },
...
      "InviteUserToRoomDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
      "CreateRoomInviteLinkDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "nullable": true,
            "deprecated": true,
            "description": "Ignored. The roomId path parameter is authoritative."
          },
          "maxUsage": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "description": "0 means unlimited uses."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateRoomInviteLinkDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "nullable": true
          },
          "maxUsage": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "description": "0 means unlimited uses."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RatedMovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dbId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "vote_average": {
            "type": "number",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "imdbRate": {
            "type": "number",
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "addedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PublicUser"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dbId",
          "title",
          "poster_path",
          "vote_average",
          "overview",
          "original_title",
          "video",
          "release_date",
          "imdbRate",
          "imdbLink",
          "isWatchedTogether",
          "createdAt",
          "updatedAt",
          "addedBy",
          "genres",
          "ratings",
          "apiVersion"
        ]
      },
      "RatingsResponse": {
        "type": "object",
        "properties": {
          "movies": {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RatedMovie"
                }
              },
              "meta": {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              "links": {
                "$ref": "#/components/schemas/PaginationLinks"
              }
            },
            "required": [
              "data",
              "meta",
              "links"
            ]
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUser"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "movies",
          "users",
          "apiVersion"
        ]
      },
      "AddedRatingResponse": {
        "type": "object",
        "properties": {
          "rate": {
            "type": "integer"
          },
          "movie": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "dbId": {
                "type": "integer"
              },
              "title": {
                "type": "string"
              },
              "poster_path": {
                "type": "string",
                "nullable": true
              },
              "vote_average": {
                "type": "number",
                "nullable": true
              },
              "overview": {
                "type": "string",
                "nullable": true
              },
              "original_title": {
                "type": "string",
                "nullable": true
              },
              "video": {
                "type": "string",
                "nullable": true
              },
              "release_date": {
                "type": "string",
                "nullable": true
              },
              "imdbRate": {
                "type": "number",
                "nullable": true
              },
              "imdbLink": {
                "type": "string",
                "nullable": true
              },
              "isWatchedTogether": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "addedBy": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PublicUser"
                  },
                  {
                    "type": "object",
                    "nullable": true,
                    "enum": [
                      null
                    ]
                  }
                ]
              },
              "genres": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            },
            "required": [
              "id",
              "dbId",
              "title",
              "poster_path",
              "vote_average",
              "overview",
              "original_title",
              "video",
              "release_date",
              "imdbRate",
              "imdbLink",
              "isWatchedTogether",
              "createdAt",
              "updatedAt",
              "addedBy"
            ]
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "rate",
          "movie",
          "apiVersion"
        ]
      },
...
      "InvitationCreatedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "id",
          "apiVersion"
        ]
      },
...
      "InvitationsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invitation"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "meta",
          "links",
          "apiVersion"
        ]
      },
...
      "CreatedInviteLinkResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "token": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "maxUsage": {
            "type": "integer"
          },
          "uses": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "room": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              }
            },
            "required": [
              "id"
            ]
          },
          "createdBy": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              }
            },
            "required": [
              "id"
            ]
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "token",
          "maxUsage",
          "uses",
          "createdAt",
          "isActive",
          "room",
          "createdBy",
          "apiVersion"
        ]
      },
...
      "InviteLinksResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InviteLink"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "meta",
          "links",
          "apiVersion"
        ]
      },
...
      "RemovedInviteLinkResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "id",
          "apiVersion"
        ]
      },
      "InvitePreviewResponse": {
        "type": "object",
        "properties": {
          "room": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "image": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "description",
              "image"
            ]
          },
          "inviter": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "avatar"
            ],
            "nullable": true
          },
          "canJoin": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "room",
          "inviter",
          "canJoin",
          "message",
          "apiVersion"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}